C-Kermit 304(dev16) 2016/01/20 Unix VMS
Mark Sapiro noticed that "OK" is no longer a valid IF condition (it was a
synonym for SUCCESS), which is true; I removed it years ago because of some
conflict but didn't make a note of it, can't remember the reason, but IF OK
still showed up in HELP IF.  Removed it in ckuus2.c, and put a new version
of the iftest script in the script library.  31 Dec 2015.

Made a new "linux" target for the makefile.  Kept the previous one available
as "linux-2015".  The feature tests in the old one looked for libraries
in specific places, like "if test -f /usr/lib64/libncurses.so.5" which is
stupid because there are an endless number of places where Linux packagers
will think up to put them, especially for multiarchitecture distributions.
Well, who really knows where the libraries are, and what the library search
path is?  ld, right?  So now instead of looking in 20 different places
for (say) libncurses under several different name, we just ask ld if it
can find it and if so -- and if we also can identify the corresponding
header file (if any) -- then we set the appropriate HAVE_blah feature flags.
makefile, 18 Jan 2016.

Updated copyright year to 2016.  ckcmai.c, 18 Jan 2016.