C-Kermit 304(dev21) 2017/04/21 Unix VMS
A development (pre-alpha, pre-beta) release.  The only siginifcant change is
a new ability to allow the FOPEN command to "open" the standard input stream
via a new command FOPEN /STDIN.  This allows Kermit scripts to read the
output of another program through a pipe, as in Unix, using its regular
FREAD command.  There are also some minor additions to the S-Expression
(LISP) command repertoire.  Dev.20 of February 16, 2016, was just a bug fix
for Mac OS X.