C-Kermit 304(dev22) 2017/05/01 Unix VMS
Improved parsing of macro arguments that are grouped versus those that
contained grouping characters as data.  Command-parsing error messages
improved somewhat but more work is needed.  Some improvements to \fsplit().
Some improvements to the handling of built-in internal macros such as IF,
FOR, and SWITCH.  The command parsing changes were actually quite massive
and fundamental, so this version was never uploaded to the Web/FTP site.