C-Kermit 305(alpha05) 2021-11-16 Unix VMS
New support for 1,500,000bps serial port speed for platforms that offer
it.  Numerous adjustments to APIs, header files, and C compilers that
changed out from under C-Kermit.  A file-transfer bug fixed that resulted in
the truncation of the filename when using ultra-short packets.  Also a
problem was fixed that occured when transferring any file that contained
0xff bytes through an external Xmodem protocol.  Most notably we have
successful builds on modern VMS operating system versions on various
hardware platforms.