C-Kermit 10.0 Beta.04, 03 June 2022
Corrects one or two syntax errors fatal to compilation.

In the (now) time-honored tradition of releasing a second Beta just
hours after the first, here is Beta.04 (the curse of the odd-numbered
Betas?).

In Beta.03, two compilation errors that I fixed repeatedly — at
least one of them fatal — somehow kept coming back. Senior confusion,
no doubt... editing or uploading the wrong copy of the file. You can
see why I'm anxious to get out of this business! The corrections for
Beta.04 are:

    ckctel.c:

    tn_ssbopt(TELOPT_NEWENVIRON,TELQUAL_SEND,(CHAR *)request,
               strlen((CHAR *)request)); /* 2022-01-27 */

    should be:

    tn_ssbopt(TELOPT_NEWENVIRON,TELQUAL_SEND,request,
               strlen((char *)request)); /* SMS 2022-06-03 */

Note EVERY makefile target should include the equivalent of gcc/clang
-funsigned-char if available. Has anybody ever actually used *signed*
chars??? (I haven't changed the makefile targets myself yet, but
readers are welcome to do this for their platforms and report the
results so I can incorporate the change where it works.)

  ckuusx.c:

    ck_curpos(int row, col) {

  should be:

    ck_curpos(int row, int col) {

At least one of these (#2) prevented successful compilation on at least
one platform (Raspberry pi OS).

Incidentally, I find myself confused about pre-ANSI function argument
list syntax. I see several forms in ckuusx.c:

         int ck_curpos(row, col) int row, col; {...}
         int ck_curpos(row, col) int row, int col; {...}
         int ck_curpos(row, col) int row; int col; {...}

Only the first one seems to be causing trouble in C-Kermit, even though
it was allowed in K&R C (see pp. 23, 24, 54, and 58 of the original 1978
C book). I see examples of (3) in the same book (e.g. pp. 60, 67, 78),
but I don't see examples of (2) in the book, and yet when when (1) caused
problems and was changed to (2) the problem went away. For Beta.04 I'm
going to err on the side of not making too many changes.

I also have a report that constructions like "*s++;" generates "Result
unused" warnings in Clang, e.g. in ckclib.c about line 3190, but I don't
believe that these are a problem. As Brian Kernighan said in 1974, "The
'*s' returns a character; the '++' increments the pointer so we'll get
the next character next time around. As you can see, as we make things
more efficient, we also make them less clear. But '*s++' is an idiom so
common that you have to know it." In this case, were toodling through a
string character by character in a loop. So the result is unused in the
first iteration but will be used in the next. The are countless instances
of this usage in the code and I doubt that the risk of changing them is
worth the benefit.