Fix handling of BEL as a terminator for OSC sequences
The Windows Console Host (when using ConPTY) uses OSC sequences terminated with a BEL to set the window title. CKW did not support this (except for when emulating aixterm but this was broken) so after receiving the set title command CKW would append all further received data to the OSC buffer - not great. This change fixes the bug that prevented this from working with aixterm and also opens it up for all terminal emulations. I hope this doesn't break anything...