Include sys/time.h
To fix errors such as:
ckcftp.c(4745): Error! E1031: Name 'tv_sec' not found in struct/union timeval
ckcftp.c(4746): Error! E1031: Name 'tv_usec' not found in struct/union timeval

Not sure if this is a watcom-specific issue or if this file is just broken for all OS/2 compilers.