Fix buffer overflow bug in OSC title code
If the title was longer than the buffer then the NULL termination could be written past the end of the buffer. Also bumped the buffer length up from 31 to 64 and switched the os2settitle function from sprintf to snprintf.

This fixes the crash bug when using PTYs in VC14 builds.