Update setenv.bat for zinc
It will now try to detect the compiler and set the appropriate zinc build configuration (if there is one), then check to see if a zinc build already exists. If it does not it outputs instructions for how to build it. If it does, it marks zinc as available alongside any other optional dependencies that were found.

mkzinc.bat no longer sets a zinc build configuration but instead relies on the one chosen by setenv.bat. It will only copy makefiles if they don't already exist and if the build fails it outputs  an error and a suggestion to run cleanall.bat. If the build succeeds it marks Zinc as available.