Negotiate best TLS available (fixes OpenSSL 3.0 compatibility)
On OpenSSL 1.1.0 or newer. On 1.0.1/1.0.2, use TLS 1.2. OpenSSL 0.9.8 and 1.0.0 will continue to use the deprecated TLS 1.0.

Previously we were always using TLS 1.0 which is usually disabled now for security reasons and not available by default in OpenSSL 3.0.x. So this not only fixes OpenSSL 3.0 support but also means the HTTP command works reliably with https sites.