Don't include SSL and other build options in the herald
The primary means of distribution for the Windows version is in binary form - everyone has the same build options so there isn't much reason to advertise them in the herald. Additionally, the windows version has a lot more build options than just SSL so not much reason to call out SSL specially.