Restore SRP support (#149)
* Restore SRP support

setenv will now detect its presence in \srp and turn SRP support on if libdes is available (as SRP support depends on k95crypt). Seems to build fine with Visual C++ 6.0 - other compilers are untested at this time.

Setenv will also now detect Kerberos for Windows and turn on kerberos support. This is entirely untested at this time.

The SRP utilties, srp-passwd an srp-tconf should also build given openssl and a sufficiently old version of SRP. They're not currently built as part of the main target due to their dependency on ancient SRP and the fact the version of tconf that comes with SRP should be fine.

* krypto.dll should be distributed too

* Add makefiles for SRP

SRP 2.1.2 is only compatible with OpenSSL prior to 1.1 currently which limits its usefulness somewhat

* Update README.md