Make the status line smarter - fixes #211
The status line will now try to make best use of available space. LEDs and help text are always shown, protocol is shown if there is room, then hostname is shown if there is room, then exit text is shown if there is room.

When there is sufficient space there is also now an extra two columns of padding between the exit text and the hostname.

The exit text can also now be up to 30 characters long (up from 20)