Look for both the dialer and connection manager
If the connection manager is found, run that - we may in the future give it the option to defer to the dialer if the user prefers. If the connection manager isn't found (which it won't be, because it doesn't work well enough yet to be anything more than a super secret project), run the dialer.