Fix #227 - SSH, SSL and Kerberos options not visible
This issue primarily affects 64bit builds. The dialer tries to load the DLLs to see if they're there but this won't work with 64bit C-Kermit where the DLLs are 64bit and the dialer remains 32bit. The ever changing name of the OpenSSL DLL makes even just checking if the file exists difficult.

Easier to just make these options permanently available.