Remove edition
I don't think it makes sense to keep tracking editions of the manual anymore - ideally it would be updated for each release of K95 that has anything more than bug fixes, so the edition number would go up with most K95 releases. Having the actual K95 version number in the manual should be enough to differentiate different versions of the manual.