Re-enable NetBIOS
Thanks to MichalN, the NetBIOS code no longer crashes on application start.

Its now working as well when as in Kermit 95 2.1.2 which is to say not at all. Looks like 1.1.17 (21 June 1998) was probably the last release where it actually worked due to the code that actually calls os2_netopen for NetBIOS connections being removed sometime between 1.1.17 and 1.1.19 (February 2000)