Revert change from K95 2.2 fixing k95stdout
Sometime after K95 2.1.3 the init code was changed to only skip starting the term update thread if K95 was starting up in server mode (iksd) and using stdout (k95stdout). This breaks all the times when k95 is starting up in normal mode using stdout as the term update thread ends up clearing the terminal, drawing the status line and clearing the screen again on exit (which renders the `-h` and `--help` options unusable).

I'm not sure why this change was originally made, and quite possibly undoing it will reintroduce some bug in iksdnt but probably a less impactful one than stdout not working properly.