Improve speed of -h and --help command line arguments
Closing the keyboard handler thread (KbdHandlerCleanup()) takes a few seconds for some reason. If we're just showing usage information and exiting, then we're not expecting any keyboard input so we can speed up the exiting bit a whole lot by not starting the keyboard handler.

There is probably a bunch of other init and cleanup stuff we could skip too, but for now this is enough to get is showing usage information in a reasonable timeframe.