Move some of the mkdocs script into REXX
The Kermit script only version of it was leaking memory somehow causing it to fail part way through. Most of the individual file processing work (searching for tags within files and replacing them, etc) is now done in REXX which can do it much faster and a bit easier thanks to regular expressions.