Fix unresolved external symbol wolfSSL_Rehandshake
WolfSSL has to be built with -DHAVE_SECURE_RENEGOTIATION for this to work, something the CMake makefiles can't currently do. And Kermit 95 doesn't seem to call ck_ssl_renegotiate_ciphers() at all anyway so we can just exclude it. Its possible that its not used anywhere ever and it can just be deleted, but I don't have the Unix or OpenVMS code handy at the moment in order to check.