Initial pass at more colors!
This *really* should have been a whole series of commits rather than one jumbo commit. The reason for the jumbo commit is that I was sort of working on all aspects of this all at once rather than one at a time, and having it all in one commit made it a bit easier for the IDE to show me before/after views so I could check for errors more easily.

This commit:
 * Changes the color attributes type from `unsigned char` to the new typedef `cell_video_attr_t`. This resulted in every single line of code dealing with color needing to be changed.
 * Replaced all instances where the contents of the color attribute was being manipulated directly (eg, shifting it right to get the BG color, ORing it to combine a FG and BG color, etc) with macros enabling the definition of cell_video_attr_t to be easily changed
 * Added a compile option to support for xterms 256-color and 88-color palettes using a 16-bit color attribute rather than the normal 8-bit. A new command (SET TERM COLOR PALETTE) is added to select a palette, and \v(color_palette) shows you the current setting, SET GUI RGBCOLOR and SHOW GUI updated to support the new palettes
 * Implements SGR 38 and 48, with RGB values quantized to the current palette. In console builds, this is always the 16-color palette.
 * Completely reworks the way operating system commands are parsed to better handle implementing a bunch of new ones
 * Implements xterms OSC-4 and OSC-104 (set and reset palette)
 * Implements xterms OSC-2 and OSC-I for setting the title
 * Implements linuxes OSC-P and OSC-R for setting and resetting the palette

Further commits will follow to implement 24-bit RGB support, and possibly the additional xterm OSCs for setting attribute-as-color values, etc.