Support disabling true bold and setting a fixed color instead
This really only affects K95G, but the functionality is available in K95 also. Why turn off true bold?
 * You don't want a bold font. Perhaps your application is using the bold SGR for color, not boldness.
 * Because xterm and teraterm let you
 * Because xterm has an OSC for setting a fixed bold color

When true bold is off and a fixed color has not been set with `SET TERM ATTR BOLD OFF COLOR`, if the foreground is a color from the aixterm16 palette its intensity bit will still be set as some applications use this to access the upper 8 colors.