Use sendchars instead of sendcharsduplex for OSC responses
I'm not really clear on what the difference between sendchars and sendcharsduplex is, and using sendcharsduplex *does* work, but that function is almost exclusively used for sending keyboard events - there is/was only one instance of it being used for something else; everything else that needed to send responses used sendchars. And if everything else was using sendchars, probably new stuff should too.