Remove NT implementation of getshiftstate()
The implementation is totally wrong - pointer to a single byte is being passed to a function that expects an array of 256 bytes.

The NT implementation is never actually used; the function is only ever called from the OS/2 mouse handling code. So the NT implementation can be safely removed. That way we won't accidentally use it in the future.