Change window size rather than terminal height
when showing or hiding the status line. This matches the behaviour of the
VT520 and plays nicer with terminfo. The change is only effective for the K95 terminal type for now as I'm not sure how the VT320 and VT420 behave. If it later turns out the VT320 and VT420 behave the same as the VT520, then this change will be applied to those emulations too.