On second thoughts, don't disable the status line by default
Having it off by default means we end up with a 25-line tty by default which is a bit non-standard. And because of the way we toggle the status line without affecting the tty height for the K95 terminal type, if we try to compensate for this by dropping the height back to 24 in settermtype() it ends up clearing the command screen because it gets resized too.

Also, the status line is required for a number of features like find, compose and bookmarks so having it off by default means those features are broken by default.

If we want to revisit this in the future and try to have it disabled by default again it will need to be done in a smarter way - probably the command screen also needs to take the same status line on/off approach when the terminal type is K95.