Implement OSC-52
This allows both querying (reading) the clipboard, and setting (writing) it, translating to/from the remote character set as required.

The feature is off by default, and can be enabled in one or both directions via SET TERMINAL CLIPBOARD-ACCESS. In K95G, when the remote host uses OSC-52 you can optionally choose to receive a notification on Windows 2000 or up.