An initial attempt at a terminfo description
Along with a kermit script to install it. Its far from perfect, but I can't really afford to spend much more time on this at the moment and it *seems* to work ok for now. If it turns out people want to use it then more improvements can always come later.