Prevent unlikely division by zero
os2settimo() divides a number of values by the connection speed so setting a speed of zero could cause problems. As far as I can tell, a speed of zero can only *currently* be set by sysinit() during program startup, and at that point in program startup it doesn't currently seem to cause problems but better to be safe and force a valid value. Cov-462528