Increased size of the connection properties dialog
While it doesn't look as nice at this larger size and its size compared to the main window is a bit off, a couple of the tabs (printer and file transfer especially) were really too cramped at the more compact size and most tabs weren't going to fit any extra settings at all on them, so this really had to be done.

With the extra space now available on the Terminal tab, the terminal type box has been enlarged and it now shows friendly names rather than terminal type keywords.