Implement the GUI Colors tab
This comes with color theme support in the storage layer with a handful of default themes defined. No GUI for creating or editing color themes though, but the storage layer does support it if we wanted to implement it someday.

The GUI Colors tab lets you pick an existing theme, or enter RGB values yourself.