Fix missing return statement in WrtNCell
In KUI builds, WrtNCell had no return statement. WrtNCell is only used for writing to the console window - something KUI builds *do not have*, so rather than adding a return statement the entire function is now excluded from KUI builds.

This also required excluding clearcmdscreen() from KUI builds (not needed, and never worked for KUI builds anyway as it relied on WrtNCell to do the clearing), and also TermScrnUpd which renders vscrns to the console window (so is never called in KUI builds).

So bug fixed, a big function that was called unnecessarily is no longer called, and a bunch of unnecessary code is no longer included in k95g.exe

Cov-462353