Initial support for Jump Lists in the Connection Manager
Its not perfect, and there are probably bugs, but it does appear to work. Its about time I got back to work on Kermit 95 proper so there probably won't be much more work on the Connection Manager for a few months, but that should give me a good opportunity to test recent changes.