Add support for setting an explicit App User Model ID
This is only done in the presence of the Kermit 95 Connection Manager, a still in-development replacement for the Dialer. While this is probably still a few years away from release, I'm using it while I intermittently develop it so its useful to have support for it in the main branch.