Do something useful for xterm private mode 1043
We cant "raise the window on bell" as Windows (for good reason) doesn't allow applications to steal focus. So instead we treat it the same as 1042 (flash the window). This matches observed behaviour on Ubuntu where  whatever window manager Ubuntu uses doesn't seem allow focus stealing either resulting in xterms attempt to take focus being treated the same as urgency