Don't preserve the intensity bit when setting non-intense colors
I don't know if the prior behaviour was a bug or not, but its not what most software expects today and its not what xterm does so its a problem for the K95 terminal type. The prior behaviour remains in place for all other terminal types.

Kermit 95 has *always* preserved the intensity bit ever since the oldest OS/2 C-Kermit release I can find. Back then, setting the bold attribute set the intensity bit too and there was no support for the aixterm extensions for setting intense colours directly. The aixterm extensions were later added in the Kermit 95 days for which I have no source history but this was likely when setting bold stopped manipulating colors directly. It *may* be a bug that SGR-30 through -37 and -40 through -47 were not updated to overwrite the intensity bit at this time, or perhaps this is exactly what aixterm did and K95 is one of the few terminal emulators to copy that behaviour. Either way, the K95 terminal type does not do this now as nothing expects it today and it causes problems.