Fix RNG seeding on the command thread
On modern Windows (all windows?) the RNG is per-thread, so you have to call srand() on each thread that's using rand(), and apparently the thread that main() runs on isn't what eventually becomes the command thread. I'm not sure it switches threads, but rand() is currently only used on the command thread for \frandom(), so we'll just track which thread the RNG was last initialised on and on \frandom() if the thread ID doesn't match we'll do it again. Fixes #483.