New workflow option to do a stable build
This will be used for non-beta/RC release builds when we eventually need to do one.

Additionally, unless the current commit has a tag the build will be marked as a development test build - all this does is add the prefix "Pre-" to the test build type (eg, "Pre-Beta 8") to indicate it isn't an actual Beta release build.

Also, the commit SHA being built is passed into the build system, though nothing is currently done with this.